requirements.txt
tensorflow
opencv-python
pyyaml